+++ pyproject.toml
@@ -33,3 +33,6 @@ pydantic = ">=2.9.2,<3"
 polars = ">=1.11.0,<2"
 pyarrow = ">=17.0.0,<18"
 tqdm = ">=4.66.6,<5"
+
+[project.entry-points.pytest11]
+property_models_fixtures = "property_models.dev_utils.fixtures"
